requirements.txt

pydantic
python-dotenv
streamlit
pandas
langchain
langchain_groq

#Google Gemini Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub